+++ src/lib/utils/options/sources/typedoc.ts
@@ -548,6 +548,7 @@ export function addTypeDocOptions(options: Pick<Options, "addDeclaration">) {
             includeCategories: false,
             includeGroups: false,
             includeFolders: true,
+            compactFolders: true,
         },
     });