+++ src/lib/models/reflections/project.ts
@@ -4,7 +4,7 @@ import { ReferenceReflection } from "./reference";
 import type { DeclarationReflection } from "./declaration";
 import type { SignatureReflection } from "./signature";
 import type { ParameterReflection } from "./parameter";
-import { IntrinsicType } from "../types";
+import { IntrinsicType, makeRecursiveVisitor, Type } from "../types";
 import type { TypeParameterReflection } from "./type-parameter";
 import { assertNever, removeIf, removeIfPresent } from "../../utils";
 import type * as ts from "typescript";
@@ -49,7 +49,7 @@ export class ProjectReflection extends ContainerReflection {
      *
      * This may be replaced with a `Map<number, Reflection>` someday.
      */
-    reflections: { [id: number]: Reflection } = {};
+    reflections: { [id: number]: Reflection } = { [this.id]: this };
 
     /**
      * The name of the package that this reflection documents according to package.json.
@@ -151,6 +151,22 @@ export class ProjectReflection extends ContainerReflection {
         if (filePath) {
             this.files.registerReflection(filePath, reflection);
         }
+    }
+
+    /**
+     * Removes references to reflections contained within the provided type.
+     * Plugins which overwrite types on reflections should pass the type to this
+     * method before overwriting the property.
+     * @since 0.26.6
+     */
+    removeTypeReflections(type: Type | undefined) {
+        type?.visit(
+            makeRecursiveVisitor({
+                reflection: (type) => {
+                    this.removeReflection(type.declaration);
+                },
+            }),
+        );
     }
 
     /**