+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Bug Fixes
 
 -   The `suppressCommentWarningsInDeclarationFiles` option now correctly ignores warnings in `.d.cts` and `.d.mts` files, #2647.
+-   Restored re-exports in the page navigation menu, #2671.
 
 ## v0.26.5 (2024-07-21)