+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 -   The `suppressCommentWarningsInDeclarationFiles` option now correctly ignores warnings in `.d.cts` and `.d.mts` files, #2647.
 -   Restored re-exports in the page navigation menu, #2671.
+-   JSON serialized projects will no longer contain reflection IDs for other projects created in the same run. Gerrit0/typedoc-plugin-zod#6.
+-   In packages mode the reflection ID counter will no longer be reset when converting projects. This previously could result in links to files not working as expected.
 
 ## v0.26.5 (2024-07-21)